Phase 5 Task 8: address CodeRabbit PR review (helper relocation, EOF banner, metadata-normalization lock)
